code cleanup and modularization

+++ routes/editAdmins.js
@@ -4,38 +4,35 @@ var router = express.Router();
 /**********************************************
 * Route for saving edits or deleting accounts.
 **********************************************/
-router.post('/', function(req, res, next) {
-	console.log(req.body);
-	if(req.body['Delete'] == "Delete"){
-	   appRepo.removeAdmin(req.body.id)
-       .then((data) => console.log('Succesfully removed admin'))
-	   .catch((error) => console.log('Error removing admin', error));
-      	appRepo.getAllAdmins().then((admins) => {
-			res.render('accountsMain',  {admins: admins, title: "Accounts" });
-     	}).catch(error => console.log('Error getting all admin: ', error));
+router.post('/', function (req, res, next) {
+    if (req.body['Delete'] == "Delete") {
+        if (req.body.id != req.session.loggedInId) {
+            appRepo.removeAdmin(req.body.id)
+               .then((data) => console.log('Succesfully removed admin'))
+                .catch((error) => console.log('Error removing admin', error));
+            var info = "Deleted Admin ID = " + req.body.id;
+        }
+        else {
+            var info = "Cannot delete logged in Admin account";
+
+        }
+        appRepo.getAllAdmins().then((admins) => {
+            res.render('getAllAdmins', { info, admins });
+        }).catch((error) => console.log('Error getting all admins', error));
      }
      else if(req.body['Edit'] == "Edit"){
-		appRepo.getAdminById(req.body.id).then((adminToEdit) => {
-   			res.render('accountsMain',  {adminToEdit: adminToEdit, title: "Accounts", subTitle: "Edit Admin" });
-	  }).catch((error) => console.log('Error editing admin', error));
+        appRepo.getAdminById(req.body.id).then((admin) => {
+            res.render('editAdmins', { admin });
+	  }).catch((error) => console.log('Error getting admin to edit', error));
 	 }
 	 else if(req.body['Save'] == "Save"){
 	 		appRepo.updateAdminNameAndEmail(req.body.id, req.body.name, req.body.email).then((admin) => {
-				console.log(admin);
 	    	 }).catch((error) => console.log('Error updating admin', error));
-		      appRepo.getAllAdmins().then((admins) => {
-				res.render('accountsMain',  {admins: admins, title: "Accounts" });
-	 	   }).catch((error) => console.log('Error updating admin', error));
+				res.redirect('getAllAdmins');
 	 }
 
      else{
-     	appRepo.updateAdminEmail(req.body.email, req.body.id)
- 	 		.then((data) => console.log('Succesfully updated admin email'))
-     	.catch((error) => console.log('Error updating admin email', error));
-
-     	appRepo.getAllAdmins().then((admins) => {
-			res.render('accountsMain',  {admins: admins, title: "Accounts" });
-     	}).catch(error => console.log('Error getting all admin: ', error));
+			res.redirect('getAllAdmins');
      }
 });
